+++ src/App.tsx
@@ -1,19 +1,19 @@
 import { Routes, Route } from 'react-router-dom';
-import Home from './components/screens/Welcome';
-import BVN from './BVN';
-import LoginScreen from './components/screens/LoginScreen';
-import OTP from './components/screens/OTP verification';
-import Validation from './components/screens/Validation';
-import Valid2 from './components/screens/ProfileValidation';
-import Valid3 from './components/screens/TermsValidation';
+import Home from './components/screens/WelcomeScreen/Welcome';
+import BVN from './components/BVN';
+import LoginScreen from './components/screens/LoginScreen/LoginScreen';
+import OTP from './components/OTP/OTP verification';
+import Validation from './components/Validation/Validation';
+import Valid2 from './components/Validation/ProfileValidation';
+import Valid3 from './components/Validation/TermsValidation';
 import TC from './components/screens/Terms'
 import Terms from './components/screens/T&C';
 import Conditions from './components/screens/Terms';
 import Done from './components/screens/Successful';
 import Protect from './components/screens/Protect'
 import NIN from './components/screens/ProtectNIN'
-import Form from './components/screens/Form';
-import OTP2 from './components/screens/OTP2';
+import Form from './components/Form/Form';
+import OTP2 from './components/OTP/OTP2';
 const App = () => {
    return (
       <>
