+++ src/components/EnterNIN/EnterNIN.tsx
