+++ src/components/OTP/OTP2.css
@@ -8,7 +8,7 @@ body {
 
  .fxemoji-bolt {
     height: 44px;
-    left: 119px;
+    left: 77px;
     position: absolute;
     top: 620px;
     width: 40px;
