--- src/components/screens/Default/Man-Frame.css
@@ -1,83 +1,0 @@
-.man-frame {
- height: 1024px;
- position: relative;
- width: 717px;
-}
-.man-frame .overlap {
- height: 884px;
- left: 28px;
- position: absolute;
- top: 59px;
- width: 649px;
-}
-.man-frame .img-lady {
-    border-radius: 189px;
-    margin: 30px;
-}
-.man-frame .group {
- height: 355px;
- left: 133px;
- position: absolute;
- top: 529px;
- width: 398px;
-}
-.man-frame .overlap-group {
- height: 355px;
- position: relative;
- width: 396px;
-}
-.man-frame .everything-is-ea-y {
- color: #ffffff;
- font-family: "Montserrat-Regular", Helvetica;
- font-size: 64px;
- font-weight: 400;
- left: 0;
- letter-spacing: 0;
- line-height: 64px;
- position: absolute;
- text-align: center;
- top: 0;
- width: 396px;
-}
-.man-frame .text-wrapper {
- line-height: 100px;
-}
-.man-frame .span {
- font-family: "Montserrat-SemiBold", Helvetica;
- font-size: 72px;
- font-weight: 600;
- line-height: 100px;
-}
-man-frame .text-wrapper-2 {
-    font-family: "Montserrat-SemiBold", Helvetica;
-    font-size: 124px;
-    font-weight: 600;
-   }
-   .man-frame .fxemoji-bolt {
-    height: 64px;
-    left: 141px;
-    position: absolute;
-    top: 123px;
-    width: 64px;
-   }
-   .man-frame .div {
-    align-items: center;
-    border-radius: 20px;
-    display: inline-flex;
-    gap: 5px;
-    left: 336px;
-    padding: 2px 3px;
-    position: absolute;
-    top: 996px;
-   }
-   .man-frame .rectangle {
-    background-color: #ffffff99;
-    border-radius: 5px;
-    height: 7px;
-    position: relative;
-    width: 28px;
-   }
-   .man-frame .icons {
-    flex: 0 0 auto;
-    position: relative;
-   }
