+++ src/components/screens/Default/OTP.css
@@ -8,7 +8,7 @@ body {
 
  .fxemoji-bolt {
     height: 44px;
-    left: 1px;
+    left: 79px;
     position: absolute;
     top: 620px;
     width: 40px;
