+++ src/components/screens/LoginScreen/style.css
@@ -74,6 +74,7 @@
   letter-spacing: 0;
   line-height: 32px;
   position: relative;
+  text-align: center;
 }
 
 .playa {
